+++ packages/assets-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add accountAssetInfo to enrich fungible balance data with per-asset snap metadata, such as Stellar trustline fields. Enrich snap balance fetches and async balance updates by calling getAccountAssetInfo and merging the returned data into asset balances ([#9217](https://github.com/MetaMask/core/pull/9217))
+
 ## [10.0.1]
 
 ### Changed